pyproject.toml
[tool.poetry]
name = "wedding-api"
version = "0.1.0"
description = ""
authors = ["Rafa Molitoris <rafa.molitoris@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
SQLAlchemy = "^2.0.20"
pydantic = "^2.3.0"
uvicorn = "^0.23.2"
pandas = "^2.0.3"
jinja2 = "^3.1.2"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
